Create label.yml
This pull request implements an automated issue labeling workflow using GitHub Actions. It triggers when issues are opened or reopened, and it scans the issue's title and body for keywords. If it finds a match, it applies the corresponding label. If no matching labels are found, it applies the "triage" label. This helps to organize issues and streamline the issue management process.


Signed-off-by: Thiago-A-Esteves <121464376+Thiago-A-Esteves@users.noreply.github.com>